rfctr(html): extract HTML-specific elements
